improve authentication query compatibility
some clients send the salt, token and password even if some of those
fields are empty. before this fix, sending the salt and token, even if
empty, ignored the password field. this caused problems if the server
was trying to use password authentication and left those other two
fields empty. now if the token and salt are empty then they are ignored.