improved download resilience
don't immediately fail a multi track download on the first track
failure. when downloading from spotify some track downloads fail
beacause a track is not available and if we were downloading an artist
then no more tracks from that artist would be downloaded after that
failure. this change keeps the first error that ocurred but keeps trying
the remaining downloads and returns that error at the end.